Refine the project , fix bugs , add G unit test setup

+++ EmployeeDatabase/include/Model/Salary.h
@@ -3,30 +3,31 @@
 #include <iostream>
 #include <map>
 #include "../Helper.h"
+#include "../Regex.h"
 #include "../DBmanage.h"
 
 class Salary {
 public:
 
     Salary() = default;
-    Salary(int id, int emp_id, float amount, float base_salary, float bonus)
+    Salary(int id, int emp_id, double amount, double base_salary, double bonus)
         : Sid(id), amount(amount), base_salary(base_salary),  bonus(bonus) {}
 
     int getId() const { return Sid; }
-    float getAmount() const { return amount; }
-    float getBaseSalary() const { return base_salary; }
-    float getBonus() const { return bonus; }
+    double getAmount() const { return amount; }
+    double getBaseSalary() const { return base_salary; }
+    double getBonus() const { return bonus; }
 
-    void setAmount(float figure) { 
+    void setAmount(double figure) {
         amount = figure;
     }
-    void setBaseSalary(const float& bSalary) { base_salary = bSalary; }
-    void setBonus(const float& figure) { bonus = figure; }
+    void setBaseSalary(const double& bSalary) { base_salary = bSalary; }
+    void setBonus(const double& figure) { bonus = figure; }
 
     void viewSalary();
     void userInputSalary();
     void insertSalary(int id) const; 
-    void increment(float , std::string);
+    void increment(double, std::string);
     void deleteSalary();
     void updateSalary();
     void action() noexcept;
@@ -34,9 +35,9 @@ public:
 private:
 
     int Sid;
-    float amount{};
-    float base_salary{};
-    float bonus{};
+    double amount{};
+    double base_salary{};
+    double bonus{};
 };
 
 #endif
